* Logs added * Charging mode selection added * fixed missing entities for device

+++ src/main/java/de/vdw/io/alpha2mqtt/services/alpha/TokenService.java
@@ -17,8 +17,10 @@ import org.javalite.http.Post;
 
 import javax.inject.Singleton;
 import java.io.IOException;
+import java.net.HttpURLConnection;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
+import java.util.Map;
 
 @RequiredArgsConstructor
 @Singleton
@@ -30,36 +32,65 @@ public class TokenService {
   private final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
 
   public String getToken() {
+    log.debug("Get token from db.");
     AlphaEssToken currentToken = Base.withDb(AlphaEssToken::findCurrentToken);
 
     if (currentToken == null) {
+      log.debug("No valid token found.");
       String url = Base.withDb(() -> AlphaEssLoadJob.getLoginJob().getUrl());
 
-      String settings = Base.withDb(() -> JsonHelper.toJsonString(AlphaEssSetting.getSettings()));
+      if (url == null) {
+        log.error("No Login url found.");
+        return null;
+      }
+
+      Map<String, String> settingMap = Base.withDb(() -> AlphaEssSetting.getSettings());
+
+      if (settingMap == null || settingMap.isEmpty()) {
+        log.error("No Login settings found.");
+        return null;
+      }
+
+      String settings = JsonHelper.toJsonString(settingMap);
 
       Post loginPost =
           Http.post(url, settings)
               .header("Accept", Constants.APPLICATION_JSON)
               .header("Content-Type", Constants.APPLICATION_JSON);
 
+      if (loginPost.responseCode() != HttpURLConnection.HTTP_OK) {
+        log.error(
+            "Unexpected response code while receiving token {}: {}",
+            loginPost.responseCode(),
+            loginPost.responseMessage());
+        return null;
+      }
+
       String loginResponse = loginPost.text();
 
       try {
         ResponseDto<TokenDto> loginResponseDto =
             objectMapper.readValue(loginResponse, new TypeReference<>() {});
+
+        log.debug("Login response: {}", loginResponse);
+
         TokenDto tokenDto = loginResponseDto.getData();
 
         LocalDateTime expirationTime =
             LocalDateTime.parse(tokenDto.getTokenCreateTime(), formatter)
                 .plusSeconds(tokenDto.getExpiresIn());
 
+        log.debug("Token extracted, expiration time calculated: {}", expirationTime);
+
         currentToken =
             Base.withDb(
                 () -> {
                   // Delete all tokens as they are expired.
+                  log.debug("Deleting all existing tokens.");
                   AlphaEssToken.deleteAll();
 
                   // Save new token.
+                  log.debug("Creating new token entry.");
                   return AlphaEssToken.create(
                       tokenDto.getAccessToken(), expirationTime, tokenDto.getRefreshTokenKey());
                 });
@@ -70,9 +101,11 @@ public class TokenService {
     }
 
     if (currentToken == null) {
+      log.error("No token created or received.");
       return null;
     }
 
+    log.debug("Return token");
     return currentToken.getToken();
   }
 }
